[MIG] Migrate letsencrypt 2.0.0 to Odoo 11.0

+++ letsencrypt/migrations/11.0.2.0.0/post-migrate.py
@@ -6,6 +6,10 @@ from odoo import api, SUPERUSER_ID
 def migrate_altnames(env):
     config = env["ir.config_parameter"]
     existing = config.search([("key", "=like", "letsencrypt.altname.%")])
+    if not existing:
+        # We may be migrating from 10.0.2.0.0, in which case
+        # letsencrypt.altnames already exists and shouldn't be clobbered.
+        return
     new_domains = "\n".join(existing.mapped("value"))
     config.set_param("letsencrypt.altnames", new_domains)
     existing.unlink()
